Measure stream process duration
This information may be useful to the user in order
to build tatistics and take them into account for any
congestion control.

For example, context durations can be measured for an specific
kind of event, and so, for example, multiply by the queue
size to get the response time for the last enqueued. The
algorithm could be keep that max timeout under a given value.

Also, example is improved to introduce the max allowed
queue size in congestion control algorithm, although that
value is hardcoded on MAX_QUEUE_SIZE macro definition.

Also, rename congesion by busyConsumers which fits better
the reality.